debugging_interview/backend/requirements.txt
Flask==2.1.2
requests==2.27.1
Flask-Cors==3.0.10 # Added for easier frontend communication
python-dotenv==0.20.0 # To load .env file
Werkzeug==2.0.3 # Pinned to a version compatible with Flask 2.1.x 